+++ py_emu_lib/src/memory/mod.rs
@@ -51,7 +51,7 @@ impl PyMemory {
         }
     }
     
-    fn len(&self) -> u16 {
+    fn len(&self) -> usize {
         self.memory.len()
     }
