+++ src/emu_lib/memory/memdevice.rs
@@ -1,6 +1,5 @@
-
 pub trait MemDevice: Send {
-    fn size(&self) -> u16;
+    fn size(&self) -> usize;
     fn read(&self, addr: u16) -> u8;
     fn write(&mut self, addr: u16, data: u8) -> Result<(), &str>;
     fn is_read_only(&self) -> bool;
@@ -9,7 +8,7 @@ pub trait MemDevice: Send {
             return Err("Can't clear read-only memory");
         }
         for i in 0..self.size() {
-            self.write(i, 0).unwrap();
+            self.write(i as u16, 0).unwrap();
         }
         Ok(())
     }
@@ -30,8 +29,8 @@ impl MemBank {
 }
 
 impl MemDevice for MemBank {
-    fn size(&self) -> u16 {
-        self.data.len() as u16
+    fn size(&self) -> usize {
+        self.data.len() 
     }
     fn read(&self, addr: u16) -> u8 {
         self.data[addr as usize]
